+++ src/Entity/Race.php
@@ -35,16 +35,16 @@ class Race
     private Collection $raceResults;
 
     /**
-     * @var Collection<int, PositionBet>
+     * @var Collection<int, RaceResultBet>
      */
-    #[ORM\OneToMany(targetEntity: PositionBet::class, mappedBy: 'race')]
-    private Collection $positionBets;
+    #[ORM\OneToMany(targetEntity: RaceResultBet::class, mappedBy: 'race')]
+    private Collection $raceResultBets;
 
     /**
-     * @var Collection<int, PunishmentPoints>
+     * @var Collection<int, PenaltyPointsAward>
      */
-    #[ORM\OneToMany(targetEntity: PunishmentPoints::class, mappedBy: 'race')]
-    private Collection $punishmentPoints;
+    #[ORM\OneToMany(targetEntity: PenaltyPointsAward::class, mappedBy: 'race')]
+    private Collection $penaltyPointsAward;
 
     #[ORM\ManyToOne(inversedBy: 'races')]
     private ?Season $season = null;
@@ -52,8 +52,8 @@ class Race
     public function __construct()
     {
         $this->raceResults = new ArrayCollection();
-        $this->positionBets = new ArrayCollection();
-        $this->punishmentPoints = new ArrayCollection();
+        $this->raceResultBets = new ArrayCollection();
+        $this->penaltyPointsAward = new ArrayCollection();
     }
 
     public function getId(): ?int
@@ -140,29 +140,29 @@ class Race
     }
 
     /**
-     * @return Collection<int, PositionBet>
+     * @return Collection<int, RaceResultBet>
      */
-    public function getPositionBets(): Collection
+    public function getRaceResultBets(): Collection
     {
-        return $this->positionBets;
+        return $this->raceResultBets;
     }
 
-    public function addPositionBet(PositionBet $positionBet): static
+    public function addRaceResultBet(RaceResultBet $raceResultBet): static
     {
-        if (!$this->positionBets->contains($positionBet)) {
-            $this->positionBets->add($positionBet);
-            $positionBet->setRace($this);
+        if (!$this->raceResultBets->contains($raceResultBet)) {
+            $this->raceResultBets->add($raceResultBet);
+            $raceResultBet->setRace($this);
         }
 
         return $this;
     }
 
-    public function removePositionBet(PositionBet $positionBet): static
+    public function removeRaceResultBet(RaceResultBet $raceResultBet): static
     {
-        if ($this->positionBets->removeElement($positionBet)) {
+        if ($this->raceResultBets->removeElement($raceResultBet)) {
             // set the owning side to null (unless already changed)
-            if ($positionBet->getRace() === $this) {
-                $positionBet->setRace(null);
+            if ($raceResultBet->getRace() === $this) {
+                $raceResultBet->setRace(null);
             }
         }
 
@@ -170,29 +170,29 @@ class Race
     }
 
     /**
-     * @return Collection<int, PunishmentPoints>
+     * @return Collection<int, PenaltyPointsAward>
      */
-    public function getPunishmentPoints(): Collection
+    public function getPenaltyPointsAward(): Collection
     {
-        return $this->punishmentPoints;
+        return $this->penaltyPointsAward;
     }
 
-    public function addPunishmentPoint(PunishmentPoints $punishmentPoint): static
+    public function addPenaltyPointsAward(PenaltyPointsAward $penaltyPointsAward): static
     {
-        if (!$this->punishmentPoints->contains($punishmentPoint)) {
-            $this->punishmentPoints->add($punishmentPoint);
-            $punishmentPoint->setRace($this);
+        if (!$this->penaltyPointsAward->contains($penaltyPointsAward)) {
+            $this->penaltyPointsAward->add($penaltyPointsAward);
+            $penaltyPointsAward->setRace($this);
         }
 
         return $this;
     }
 
-    public function removePunishmentPoint(PunishmentPoints $punishmentPoint): static
+    public function removePenaltyPointsAward(PenaltyPointsAward $penaltyPointsAward): static
     {
-        if ($this->punishmentPoints->removeElement($punishmentPoint)) {
+        if ($this->penaltyPointsAward->removeElement($penaltyPointsAward)) {
             // set the owning side to null (unless already changed)
-            if ($punishmentPoint->getRace() === $this) {
-                $punishmentPoint->setRace(null);
+            if ($penaltyPointsAward->getRace() === $this) {
+                $penaltyPointsAward->setRace(null);
             }
         }