+++ src/Entity/User.php
@@ -34,16 +34,16 @@ class User implements UserInterface, PasswordAuthenticatedUserInterface
     private ?string $password = null;
 
     /**
-     * @var Collection<int, PositionBet>
+     * @var Collection<int, RaceResultBet>
      */
-    #[ORM\OneToMany(targetEntity: PositionBet::class, mappedBy: 'user')]
-    private Collection $positionBets;
+    #[ORM\OneToMany(targetEntity: RaceResultBet::class, mappedBy: 'user')]
+    private Collection $raceResultBets;
 
     /**
-     * @var Collection<int, PunishmentPoints>
+     * @var Collection<int, PenaltyPointsAward>
      */
-    #[ORM\OneToMany(targetEntity: PunishmentPoints::class, mappedBy: 'user')]
-    private Collection $punishmentPoints;
+    #[ORM\OneToMany(targetEntity: PenaltyPointsAward::class, mappedBy: 'user')]
+    private Collection $penaltyPointsAward;
 
     /**
      * @var Collection<int, WorldChampionBet>
@@ -62,8 +62,8 @@ class User implements UserInterface, PasswordAuthenticatedUserInterface
 
     public function __construct()
     {
-        $this->positionBets = new ArrayCollection();
-        $this->punishmentPoints = new ArrayCollection();
+        $this->raceResultBets = new ArrayCollection();
+        $this->penaltyPointsAward = new ArrayCollection();
         $this->worldChampionBets = new ArrayCollection();
     }
 
@@ -146,29 +146,29 @@ class User implements UserInterface, PasswordAuthenticatedUserInterface
     }
 
     /**
-     * @return Collection<int, PositionBet>
+     * @return Collection<int, RaceResultBet>
      */
-    public function getPositionBets(): Collection
+    public function getRaceResultBets(): Collection
     {
-        return $this->positionBets;
+        return $this->raceResultBets;
     }
 
-    public function addPositionBet(PositionBet $positionBet): static
+    public function addRaceResultBet(RaceResultBet $raceResultBet): static
     {
-        if (!$this->positionBets->contains($positionBet)) {
-            $this->positionBets->add($positionBet);
-            $positionBet->setUser($this);
+        if (!$this->raceResultBets->contains($raceResultBet)) {
+            $this->raceResultBets->add($raceResultBet);
+            $raceResultBet->setUser($this);
         }
 
         return $this;
     }
 
-    public function removePositionBet(PositionBet $positionBet): static
+    public function removeRaceResultBet(RaceResultBet $raceResultBet): static
     {
-        if ($this->positionBets->removeElement($positionBet)) {
+        if ($this->raceResultBets->removeElement($raceResultBet)) {
             // set the owning side to null (unless already changed)
-            if ($positionBet->getUser() === $this) {
-                $positionBet->setUser(null);
+            if ($raceResultBet->getUser() === $this) {
+                $raceResultBet->setUser(null);
             }
         }
 
@@ -176,29 +176,29 @@ class User implements UserInterface, PasswordAuthenticatedUserInterface
     }
 
     /**
-     * @return Collection<int, PunishmentPoints>
+     * @return Collection<int, PenaltyPointsAward>
      */
-    public function getPunishmentPoints(): Collection
+    public function getPenaltyPointsAward(): Collection
     {
-        return $this->punishmentPoints;
+        return $this->penaltyPointsAward;
     }
 
-    public function addPunishmentPoint(PunishmentPoints $punishmentPoint): static
+    public function addPenaltyPointsAward(PenaltyPointsAward $penaltyPointsAward): static
     {
-        if (!$this->punishmentPoints->contains($punishmentPoint)) {
-            $this->punishmentPoints->add($punishmentPoint);
-            $punishmentPoint->setUser($this);
+        if (!$this->penaltyPointsAward->contains($penaltyPointsAward)) {
+            $this->penaltyPointsAward->add($penaltyPointsAward);
+            $penaltyPointsAward->setUser($this);
         }
 
         return $this;
     }
 
-    public function removePunishmentPoint(PunishmentPoints $punishmentPoint): static
+    public function removePenaltyPointsAward(PenaltyPointsAward $penaltyPointsAward): static
     {
-        if ($this->punishmentPoints->removeElement($punishmentPoint)) {
+        if ($this->penaltyPointsAward->removeElement($penaltyPointsAward)) {
             // set the owning side to null (unless already changed)
-            if ($punishmentPoint->getUser() === $this) {
-                $punishmentPoint->setUser(null);
+            if ($penaltyPointsAward->getUser() === $this) {
+                $penaltyPointsAward->setUser(null);
             }
         }