+++ src/Service/ToastFactory.php
@@ -6,6 +6,14 @@ use App\Dto\ToastDto;
 
 class ToastFactory
 {
+    public static function generateCustomSuccessToast(string $text) {
+        $toast = new ToastDto();
+        $toast->setText($text);
+        $toast->setState(ToastDto::STATE_SUCCESS);
+
+        return $toast;
+    }
+
     public static function generateSaveSuccessfulToast() {
         $toast = new ToastDto();
         $toast->setText('Speichern erfolgreich');